moves v2 transformers to legacy->two folder
