+++ app/Http/Controllers/Web/PostController.php
@@ -5,7 +5,7 @@ namespace Rogue\Http\Controllers\Web;
 use Rogue\Services\Fastly;
 use Rogue\Services\PostService;
 use Rogue\Repositories\SignupRepository;
-use Rogue\Http\Transformers\PostTransformer;
+use Rogue\Http\Transformers\Legacy\Two\PostTransformer;
 use Rogue\Http\Controllers\Traits\PostRequests;
 use Rogue\Http\Controllers\Controller;
 
@@ -35,7 +35,7 @@ class PostController extends Controller
     protected $signups;
 
     /**
-     * @var \League\Fractal\TransformerAbstract;
+     * @var Rogue\Http\Transformers\Legacy\Two\PostTransformer;
      */
     protected $transformer;
