+++ app/Http/Transformers/Legacy/Two/ReportbackTransformer.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Rogue\Http\Transformers;
+namespace Rogue\Http\Transformers\Legacy\Two;
 
 use Rogue\Models\Post;
 use League\Fractal\TransformerAbstract;
