+++ app/Http/Transformers/Legacy/Two/SignupTransformer.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Rogue\Http\Transformers;
+namespace Rogue\Http\Transformers\Legacy\Two;
 
 use Rogue\Models\Signup;
 use Rogue\Services\Registrar;
